Add FunctionArmFeaturizer implementation and update exports in __init__.py
