+++ bayesianbandits/featurizers/__init__.py
@@ -1,9 +1,11 @@
 """Built-in arm featurizers for common use cases."""
 
 from ._continuous import ContinuousArmFeaturizer
+from ._function import FunctionArmFeaturizer
 from ._one_hot import OneHotArmFeaturizer
 
 __all__ = [
     "ContinuousArmFeaturizer",
+    "FunctionArmFeaturizer",
     "OneHotArmFeaturizer",
 ]
